requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
playwright>=1.40.0
